ultimos cambios

+++ src/Atributo.java
@@ -1,9 +1,9 @@
 
 public class Atributo {
 	private String nombre;
-	private int valor;
+	private double valor;
 	
-	public Atributo(String nombre, int valor) {
+	public Atributo(String nombre, double valor) {
 		this.nombre = nombre;
 		this.valor = valor;
 	}
@@ -35,11 +35,11 @@ public class Atributo {
 		this.nombre = nombre;
 	}
 
-	public int getValor() {
+	public double getValor() {
 		return valor;
 	}
 
-	public void setValor(int valor) {
+	public void setValor(double valor) {
 		this.valor = valor;
 	}
