+++ src/PocimaIncrementadora.java
@@ -1,23 +1,23 @@
 
 public class PocimaIncrementadora extends Pocima {
-	private int valor;
+	private double valor;
 	
-	public PocimaIncrementadora(String nombre, int valor) {
+	public PocimaIncrementadora(String nombre, double valor) {
 		super(nombre);
 		this.valor = valor;
 	}
 
 	@Override
-	public int modificarValor(Atributo atributo) {
-		int valorAtributo = atributo.getValor();
+	public double modificarValor(Atributo atributo) {
+		double valorAtributo = atributo.getValor();
 		return valorAtributo += valorAtributo * this.getValor() / 100;
 	}
 
-	public int getValor() {
+	public double getValor() {
 		return valor;
 	}
 
-	public void setValor(int valor) {
+	public void setValor(double valor) {
 		this.valor = valor;
 	}
